test: add and remove target repository tests
